--- src/advanced/components/AddButton.tsx
@@ -1,14 +1,0 @@
-interface AddButtonProps {
-  onClick: () => void;
-}
-
-export const AddButton = ({ onClick }: AddButtonProps) => {
-  return (
-    <button
-      className='rounded bg-blue-500 px-4 py-2 text-white'
-      onClick={onClick}
-    >
-      추가
-    </button>
-  );
-};
